Move Ui.java to ekud.ui package
