+++ src/main/java/ekud/commands/ExitCommand.java
@@ -2,7 +2,7 @@ package ekud.commands;
 
 import ekud.components.Storage;
 import ekud.components.TaskList;
-import ekud.components.Ui;
+import ekud.ui.Ui;
 
 /**
  * Represents a {@link Command} to terminate EKuD.
