+++ src/main/java/ekud/task/Task.java
@@ -3,8 +3,8 @@ package ekud.task;
 import java.util.HashMap;
 
 import ekud.components.TaskList;
-import ekud.components.Ui;
 import ekud.exceptions.EkudException;
+import ekud.ui.Ui;
 
 /**
  * Represents a task. Each task minimally has a description and completion status.
